SetLoggingChannelCommand -> LoggingCommand
A new command that has sub commands to set and remove the logging channel.

This will later become a command to also toggle logs, and other logging configs.
